+++ __mocks__/electron/ipcMain.ts
@@ -3,7 +3,3 @@ import { EventEmitter } from "events";
 class IpcMain extends EventEmitter {}
 
 export const ipcMain = new IpcMain();
-
-export const resetIpcMain = () => {
-  ipcMain.removeAllListeners();
-};
