+++ __mocks__/electron/ipcRenderer.ts
@@ -6,7 +6,3 @@ class IpcRenderer extends EventEmitter {
 }
 
 export const ipcRenderer = new IpcRenderer();
-
-export const resetIpcRenderer = () => {
-  ipcRenderer.removeAllListeners();
-};
