+++ src/renderer/App/AppBar/SearchBox/tests/SearchBox.test.ts
@@ -1,6 +1,6 @@
 import { screen } from "@testing-library/dom";
 import { fireEvent } from "@testing-library/react";
-import { initializeElectronCode } from "_m/startMainProcess.testHelpers";
+import { initializeElectronCode } from "_tu/initializeElectronCode";
 import { findByTestId } from "_tu/findByTestId";
 import * as blockFixtures from "_tu/fixtures/blockFixtures";
 import { renderAppWithStore } from "_tu/renderAppWithStore";
