+++ src/renderer/App/RpcIssueDialog/tests/RpcIssueDialog.cookieDialog.test.ts
@@ -1,6 +1,6 @@
 import { wait } from "@testing-library/dom";
 import { fireEvent } from "@testing-library/react";
-import { initializeElectronCode } from "_m/startMainProcess.testHelpers";
+import { initializeElectronCode } from "_tu/initializeElectronCode";
 import * as blockFixtures from "_tu/fixtures/blockFixtures";
 import { renderAppWithStore } from "_tu/renderAppWithStore";
 import { userEvent } from "_tu/smallUtils";
