+++ src/renderer/App/RpcIssueDialog/tests/RpcIssueDialog.missingBitcoinConf.test.ts
@@ -1,5 +1,5 @@
 import { fireEvent } from "@testing-library/react";
-import { initializeElectronCode } from "_m/startMainProcess.testHelpers";
+import { initializeElectronCode } from "_tu/initializeElectronCode";
 import { findByTestId } from "_tu/findByTestId";
 import * as blockFixtures from "_tu/fixtures/blockFixtures";
 import { renderAppWithStore } from "_tu/renderAppWithStore";
