+++ src/testUtils/startMockRpcServer.ts
@@ -2,7 +2,7 @@
 import { matches } from "lodash";
 import nock from "nock";
 import { NODE_ERROR } from "_c/constants";
-import { SERVER_URL } from "_m/startMainProcess.testHelpers";
+import { SERVER_URL } from "_tu/initializeElectronCode";
 import { Block, RawTransaction } from "_t/RpcResponses";
 import { blockchainInfoFixture1 } from "_tu/fixtures/blockchainInfoFixtures";
 import * as blockFixtures from "_tu/fixtures/blockFixtures";
